modify break and continue statements
