+++ .gitignore
@@ -1,0 +1,6 @@
+/*
+!.gitignore
+!*.c
+!*.h
+!readme.md
+!test3.txt
